requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
lxml==4.5.2
neobolt==1.7.17
neotime==1.7.4
parsel==1.6.0
prompt-toolkit==2.0.10
Protego==0.1.16
psycopg2==2.8.5
py2neo==4.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-arango==5.4.0
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
requests-toolbelt==0.9.1
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
urllib3==1.24.3
w3lib==1.22.0
wcwidth==0.2.5
wincertstore==0.2
zope.interface==5.1.0
